Remove INT_MIN overflow check

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: bgazur <bgazur@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 11:07:12 by bgazur            #+#    #+#             */
-/*   Updated: 2025/04/20 14:32:09 by bgazur           ###   ########.fr       */
+/*   Updated: 2025/04/22 14:19:49 by bgazur           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,7 @@
  * Skips the whitespaces (as defined in ft_isspace()).
  * Single occurence of '+' or '-' is allowed and stored (sign = +1 by default).
  * Converts and stores the number sequence in the 'result' variable.
- * Variable 'result' is of type long long int to handle INT_MAX + 1 -> INT_MIN.
- * No other overflow or underflow checks, in line with the atoi() function. 
+ * No overflow or underflow checks, in line with the atoi() function. 
  * 
  * Return value:
  * Returns the value in 'result' * 'sign' as an int (0 on error). */
@@ -37,9 +36,9 @@ static int	ft_isspace(int c)
 
 int	ft_atoi(const char *nptr)
 {
-	long long int	sign;
-	long long int	result;
-	size_t			i;	
+	int		sign;
+	int		result;
+	size_t	i;	
 
 	sign = 1;
 	result = 0;
@@ -57,5 +56,5 @@ int	ft_atoi(const char *nptr)
 		result = result * 10 + (nptr[i] - '0');
 		i++;
 	}
-	return ((int)(result * sign));
+	return (result * sign);
 }
